backend/requirements.txt
# Core Django packages
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# CORS handling
django-cors-headers==4.3.1

# Environment variables
python-decouple==3.8

# Image handling
Pillow==10.2.0

# PDF generation
reportlab==4.0.9

# Filtering
django-filter==23.5

# Development tools
black==24.1.1
flake8==7.0.0

# Testing
pytest==7.4.4
pytest-django==4.7.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0